+++ analytics_on_fhir/mii_conditions_labs.py
@@ -228,6 +228,9 @@ class PyRateQuery:
             lab_df = pd.concat(all_labs, ignore_index=True)
             if "lab_codeableconcept_code" not in lab_df.columns:
                 lab_df["lab_codeableconcept_code"] = None
+            # parse potentially problematic columns explicitly as strings
+            lab_df["lab_codeableconcept_code"] = lab_df["lab_codeableconcept_code"].astype(str)
+            lab_df["lab_quantity_unit"] = lab_df["lab_quantity_unit"].astype(str)
             lab_df.to_csv(
                 os.path.join(self.output_dir, "df_mii_labs" + suffix + ".csv"),
                 index=False,